remove some duplications on add/replace
